Major Refactoring:
The entry point of the application is now Main.java in the views
package.  I was forced to do this due to the nature of the fxmlLoader.

Got rid of the controllers package and now have all the view controllers
in the views package with the fxml files.  Again this is due to the
nature of the fxmlLoader.

I currently have a very basic example thus far that is capable of
allowing a user to enter reading values and create a Reading object from
the subsequent entries.  It also supports switching views via the
ViewNavigator.java class in the views package, which basically just
swaps nodes in and out of the root node of the main application scene.
As of now I only have a basic inputReading view.

Removed the item_id property from the Reading,java class as well as
fixed the JSON file to remove the capital "R"s and removed the item_id
from there as well.