Additional Functionality & Bug Fixes:
Fixed an error in the Reading.java constructor.  I had the readingId and
readingValue flip-flopped so the readingID would be readingValue and
vice versa.  It is now corrected.

Added a basic Writer class.  It currently writes a list of Readings to a
file named test_data_out.json.  This class is not nearly finished but it
is a good start.

Check test class for new tests.