Whoa! Lots of good stuff!!!!
Finished the Reading class Hierarchy.

Finished the jsonProcessor, which is used to import readings from a
.json file and write readings to a .json file.

The Reading class is abstract and so you cannot directly instantiate it.
You have to use the ReadingFactory.get() static method and tell it what
kind of reading you are interested in.

Refactored Main.java to be ClinicalTrialClient.java as that is our
application.

Deleted a whole bunch of uneeded classes <see commit for details>

Refactored all of the Reading classes to the reading package.

Added a JsonProcessorTest JUnit test. It is just a single test that
reads the test data json file into a list of readings and then
immediately writes the list of readings back out to a json file in the
same directory as the test data. This test should always pass!