Major UI Overhaul:
Note: I altered the Trial, and Patient classes slightly to permit me to
test some things with the UI.  The code that is in ClinicalTrialView and
ClinicalTrialViewModel is solely for example purposes and will be
updated when the other classes are fully implemented.

The UI is, at this point, one single view.  I consolidated the different
user actions right into the main view showing the list of patients and
their respective readings.  The UI is now following proper MVVM
conventions in that I moved the mixture of view and model-view logic to
their respective classes thereby decoupling the view from the model
itself, leaving the model-view responsible for interacting with the
model and presenting changes to the view.

The ViewNavigator is now removed, due to the above conventions and is no
longer needed.

All classes pertaining to the UI are now in a the package:
edu.metrostate.ics372.thatgroup.clinicaltrial.views