Almost done refactoring. Hopefully I will have everything refactored before tonight's class and then we can dole out any remaining tasks for those that have time to do it.
Anyway, with this update, you can import / export patients and readings
in to a
trial. You can add patients to a trial.

Still can't do the following:

1. Edit / Update a Patient
2. Add / Edit / Delete Readings
3. The ReadingsView needs work to display the readings correctly in the
table and make it sortable. Also needs to support adding listeners to
the selection model for change notification. so that the ReadingView can
load the reading when the selection changes.
4. The Application does not currently save it state and it needs to be
able to do that or else any entered / imported data is just lost.
5. The edit box on the Patients view doesn't currently do anything.
Would be nice if this could be used to filter the list of displayed
items in the ListView.
6. JavaDoc needs to be added to every public class / method / and
instance variable. Only the public & protected methods and variables
need JavaDoc. Simply stating a parameter name in a method is not
acceptable JavaDoc. Every parameter, return value, method name and
variable needs to be documented such that someone else using it doesn't
need to ask anyone any questions.
7. I'm sure there are more but I will add them once I recall what they
are.

The guts should be good enough that everyone can pick-up a task and run
with it. If there are not, let me know and we will make it good enough.