Small bug fix:
Fixed a small issue while adding a new patient.  The special char error
message was popping whenever a new patient was added to the trial.  It
was due to the fact that when a new patient was added the text field
clears thus changing its value to "" and an empty value is not part of
the regex pattern.  I have included an additional condition in the text
field text property listener to account for that.