Huge update!
Everything is working with the data layer now. Well, there is a bug in
the patients view / add reading view in that the buttons don't seem to
behave correctly. Need to debug to find out what is going wrong.

Still lots of work to be done but we will do that this weekend on
Saturday.