Fixed a bug in the reading view that prevented entering valid readings
The bug had to do with entering a reading for a patient from a earlier
time.

I added a hasValidDate() and hasValidDateTime() methods to the validator
to ensure that these conditions are handled appropriately.

So now you cannot add a reading to a patient prior to them starting
the trial and you cannot enter a reading for a future date and time.