Added import and export interface and support to import XML data.
Added the TrailDataImporter and TrialDataExporter interfaces.
Implemented the TrialDataXmlImporter which imports the new XML file
format that we need to support.

Add a unit test for the TrialDataXmlImporter.

Need to debug TrialDataXmlImporter and fix errors.

Need to convert JsonProcessor over to the new import / export interface.

Need to update the UI to allow for importing XML data.