Refactored the JsonProcessor to use the new Import / Export interfaces
This update affects Issue #1, Issue #3, Issue #4, and Issue #11.

The new Importer and Exporter interfaces are now cemented in to place.

The new TrialDataImporter interface is implemented by
TrialDataXmlImporter and by TrialDataJsonImportExporter.

The new TrialDataExporter interface is implemented by
TrialDataJsonImportExporter.

The ClinicalTrialView has been updated to use the
TrialDataImportExporterFactory to obtain a reference to an appropriate
TrialImporter or TrialExporter based on the type of data being imported
or exported.

Also updated the Clinic bean with proper equals and hashCode
implementations and updated the toString method.

Cleaned up the ClinicsView and AddClinicsView.