Huge refactor of the ReadingView and the ClinicsView and PatientsView.
Removed the AddPatientView and AddClinicView and merged them in to
PatientsView and ClinicsView respectively.

Added the use of Executors for kicking off the import and export jobs.

Added "clear" buttons to the Clinics and Patients views to clear the
forms the current selection.

Added the ability to update Clinic names.

Things are working pretty good.