Woot. Created additional abstraction on the SQLite datalayer.
This was needed to support Android.

Also added Migration scripts to migrate any old and new databases
to the new schema with patient status table.

updated patient bean with the status field.

need to update the ClinicalTrialStatements and procedures to account
for the new patient status table and the status field.

Need to add ability to the catalog to get the list of patient statuses

In order to add that, we first need to add a PatientStatus bean. This
simply will be a small POJO that has a string id and string
displayStatus that is used in look-up lists to allow the user to select
a new status for the patient.


Once we have that done and tested, then we can build the jar and
start the Android part of this.

If time permits and any actually cares, after we are done with android
the JavaFX UI should be updated to use the new ID field.