Updated PropertyChangeSupport in all of the model classes.
Added the ability to remove a registered listener.

Added property change support to all of the reading beans and their
related values.

Updated the views so now the Patient Status field is updated when
start and ending a patient in a trial. Withdrawn and Failed are not
an option in the desktop version at the moment.